Snekbox: remove code template
Removing the template allows for line numbers in tracebacks to
correspond to the input code.

Since stderr has been merged into stdout, exceptions will already be
captured. Thus, it is redundant to wrap the code in a try-except.

Importing of the site module has been re-enabled on Snekbox which
automatically adds site-packages to sys.path. Thus, the virtual
environment does not need to be activated anymore in a template.