Add API response dict to ResponseCodeError
The ReponseCodeError held a reference to `aiohttp.ResonseObject` to
make sure the response data was available. However, the response data
is not actually included in the Response Object, but needs to be
awaited.

Unfortunately, the ResponseCodeError is usually inspected after the
connection has been closed, triggering a ClientConnectionError when the
data was retrieved.

I solved this by adding the awaited reponse data directly to our custom
exception by awaiting the response.json() before raising the exception.