Improvements to the Wolfram cog.
This commit adds clearer error messages for all the errors
produced by the Wolfram cog if the API key is missing
in the config, or if the key isn't valid anymore.

It also humanizes the timedelta returned in the error
users get when they run out their cooldown. Instead of
telling them they need to wait 84000 seconds, it will now
tell them they need to wait 23 hours, 59 minutes ...