Make ResponseErrorCode handle non-JSON response
The previous commit assumed the API respone to always be JSON. This
leads to issues when that is not the case, such as when the API is
completely unreachable (text/html 404 response).

Updated the ResponseErrorCode exception to account for that and
updated the moderation util `post_infraction` to reflect that.