Nuking Roles.devops and the Deployment cog.
The Deployment cog does not work in our new server environment,
nor are we particularly inclined to make it work (for security
reasons). For this reason, I've removed it.

I've also removed all usages of Roles.devops, replacing them with
Roles.core_developer whenever this made sense to do, such as in
the Cogs cog.