Improve handling of long deleted messsages
The `modlog` cog failed on long messages with attachments, since the
inclusion of attachment metadata would bring the length of the embed
description to go over the character limit of 2048.

To fix this, I moved the addition of the attachment metadata to earlier
in the code so it is taken into account for the character limit.

In addition to this, I changed the truncation behavior. Instead of just
truncating the message if it's too long, we now truncate and upload the
full message to the `deleted messages` endpoint so the full message is
still available. A link to the log will be included in the message-log
embed.