Fix deleting already deleted message in antispam
Since we're in an async context, it can happen that a message was
already deleted before the antispam cog could get to it. To prevent
processing from stopping dead because of a NotFound exception, I added
a try-except and log message to handle that

In addition, corrected a small mistake: trying to indice a single
Message object. Corrected.