Resolves a breaking bug in the Dockerfile.
We were using a pipenv run script to launch the bot,
but pipenv run scripts assume that the run command will
be run inside of a virtual environment. So, the default
behaviour when we try to use a run command and no venv
exists is to create a venv.

Because we were installing all the packages to the local
environment by passing the '--system' flag to our install,
this would make the bot fail with ImportErrors.

This commit fixes it so that the Dockerfile will run the
bot using the system Python instead of the pipenv run script.