Update discord.py version to 1.2.3
I have updated the discord.py version to 1.2.3. This includes changes
throughout the entire code base, including:

- All cogs now inherit from `discord.ext.commands.Cog`;

- All of our ABCs now have `bot.utils.CogABCMeta` as a metaclass;

- All event handlers in cogs are now decorated with `Cog.listener()`;

- Some method names have changes, including:

  - get_message => fetch_message
  - get_webhook_info => fetch_webhook

- A few occurences of `get_channel` have been replaced by the new
  coroutine `fetch_channel`;

- I've had to patch a few lines of code to account for small
  differences between the versions, like `on_member_update` attribute
  names in ModLog and the fact the way we used `Context.invoke` a couple
  of times has stopped working.

In addition, I've added a patch for a bug in discord.py (with the help
of @Scragly). This discord.py version has a bug which causes the edited
timestamp not to be processed for edited messages. It's already fixed
on GitHub, but a bug fix release has not been released to PyPI. In the
meantime, I've added a patch in `bot.patches.message_edited_at` and
included conditional loading in `__main__`.

Finally, I noticed a small bug in `bot.cogs.filtering` that I fixed; I
replaces `return` with `continue` to make sure filtering for edited
messages doesn't stop after the `rich_embed_filter`.