Escape markdown in codeblock suggestion embed
https://github.com/python-discord/bot/issues/434

If the content we prepare for the codeblock suggestion embed defined
in the `bot` cog contains markdown characters (e.g., `__`, `**`), this
will cause Discord to apply markdown, since both the codeblock syntax
example as well as the codeblock result example will contain the
characters, making it a matched formatting pair.

This will hide those characters, which often have a function in the
code, and break the example of a formatted codeblock.

To deal with that, I've added a regex substitution that substitutes
markdown characters by an escaped version of that markdown character.

Example:
`'__hello__'` will become `'\_\_hello\_\_'`

I've added this substitution to both paths in the `on_message` event
listener, since we can't substitute at an earlier place due to it
generating `SyntaxErrors` with the `ast.parse` check.

This closes #434