Fix date formatting bug in infraction search
The infraction search feature did not work because of a small bug with
the date formatting: `datetime.fromisoformat` does not like the Z at
the end of the datestring the database sends back, so we need to chop
it off. I've applied the same method for doing that as already in use
in other parts of the bot codebase.