Make `!tags edit` use PATCH to update tags
https://github.com/python-discord/bot/issues/474

This commit makes sure we're using the PATCH method to update tags,
instead of the POST method. The problem was that the `!tags edit`
command was an alias of `!tags set`, which uses the POST method to
post new tags to the API.

However, when a tag with a given name already exists, the API will
refuse a POST request in an attempt to update it; we need to use the
PATCH method to the `bot/tags/{tag_name}` endpoint instead.

I fixed this by creating a separate subcommand, `!tags edit`, that
uses the correct PATCH method to update an existing tag.

This commit closes #474