Fix `cog_unload` bug in WatchChannel ABC
https://github.com/python-discord/bot/issues/482

There was small bug in the `cog_unload` method of the WatchChannel
ABC in `bot.cogs.watchchannels.watchchannel`. The problem was that it
tries to check if the Task assigned to `self._consume_task` is done
by accessing its `done` method. However, if a watch channel has not
yet relayed messages after the bot has started, it will not have a
consumption task yet, meaning this `_consume_task` attribute will be
assigned to `None`.

The solution is to change the `if` condition to:

  `if self._consume_task and not self._consume_task.done():`

This commit closes #482