Set bot as actor of antispam infractions
As mentioned in #476, the bot currently sets the actor of infractions
applied due to an antispam rule trigger to the offending member. The
reason is that we get a `Context` object from the message that
triggered the antispam rule, which was sent by the offender.

I've changed it by patching both available author attributes,
`Context.author` and `Context.message.author` with the bot user.