Show infraction reason when the bot is the actor
https://github.com/python-discord/bot/issues/476

We recently decided to hide the reason in the confirmation message
the bot sends after applying an infraction. In most situations, this
makes sense, since the message containing the invocation command
already contains the reason.

However, if the infraction was triggered by the bot itself (e.g., an
antispam trigger), this means that we're missing information that
provides context to the infraction.

This commit adds the reason back to the confirmation message, but
only if the actor of the infraction was the bot itself.

Closes #476