Ensure display name changes are logged
https://github.com/python-discord/bot/issues/489

Recently, we discovered that not all display name changes were logged
to the #user-log channel. This problem was caused by the `old_value`
or the `new_value` showing up as `None` when a user sets or removes a
guild-specific nickname. Since we ignore changes where one of the two
values is `None`, we did not log these `None->nick` or `nick->None`
events.

Since we are mainly interested in the display name of the user, and
the display name is equal to the user's guild-specific nickname if
they have set one and otherwise their username, I made the following
changes:

- Add logging of changes in the display names of members.

- Ignore nick-specific changes completely, since these changes are
  already captured by the changes in the display name we now log.

This closes #489