InfractionScheduler: create generic function for reapplying infractions
reapply_infraction() is a generic function which reapplies a given
infraction if it is still active or deactivates it if less than 60
seconds remain. Most of the code from Infractions.on_member_joined now
lives in this new function.

This simplifies the "fancier" DM and mod log that were previously sent
for superstar in on_member_join.