Autodelete offensive messages after one week.
If the filter cog filter a message that's considered as offensive
(filter["offensive_msg"] is True), the cog create a new
offensive message object in the bot db with a delete_date of one
week after it was sent.
A background task run every day, pull up a list of message to
delete, find them back, and delete them.