Add duck_pond cog.
This cog will listen for duck reactions on any message, and then:
- If the reaction was added by a staff member
- and the reaction was a duck
- and the message has not already been added to the #duck-pond

It will add the message to the #duck-pond and then add a green checkbox
to the original message to indicate that the message has been ponded.

Messages are added to the #duck-pond via webhook, so that they can
retain the appearance of having their original authors.

Once this checkmark has been added, the message will not be processed in
the future.

If the checkmark is removed and there are more than ducks_required ducks
on the message, the bot will automatically add the checkmark back.

However, if all reactions are removed, the bot does not have a
countermeasure for this. In order to implement a countermeasure, it
would be necessary to involve the API and the database.