Enhance Reddit webhook embeds
https://github.com/python-discord/bot/issues/634

I have changes the appearance of the embed generated for the reddit
webhook. The changes:

- Bold markdown around the links to prevent it breaking on android.

- Stylized the meta-data line with newly created emoji-based icons.

- Removed redundant mentions of the subreddit from the embed itself.

The emojis were uploaded to the `Emojis II` guild and the IDs have
been added to the constants files. In addition, I've set the reddit
logo as the default avatar for the webhook in the guild.

This commit closes #634