Use 'local' emojis for reddit webhook embed
Apparently, webhooks cannot use external emojis. For now, I have
added the emojis to the guild directly and updated the IDs in this
commit.

I don't really like having "utility" emojis on the main guild, but
the alternative would be to not use webhooks, which is a change we
will need to discuss before we apply it. (I don't see anything
against it now we just send a daily digest, although, in principle,
this is precisely what a webhook is meant for.)