Check only for bot's green checkmark in DuckPond
Previously, the presence of any green checkmark as a reaction would
prevent a message from being relayed to the duck pond, regardless of
the actor of that reaction. Since we only want to check if the bot
has already processed this message, we should check for a checkmark
added by the bot.

This commit adds such a user check.