Move payload checks to start of DuckPond.on_raw_reaction_add
The `DuckPond.on_raw_message_add` event listener makes an API call to
fetch the message the reaction was added to. However, we don't need
to fetch the message if the reaction that was added is not relevant
to the duck pond. To prevent such unnecessary API calls, I have moved
the code that checks for the relevance of the reaction event to
before the code that fetches the message.