Add `return_value` support and assertions to AsyncIteratorMock
The AsyncIteratorMock included in Python 3.8 will work similarly to
the mocks of callabes. This means that it allows you to set the items
it will yield using the `return_value` attribute. It will also have
support for the common Mock-specific assertions.

This commit introduces some backports of those features in a slightly
simplified way to make the transition to Python 3.8 easier in the
future.