Make post_infraction try to `post_user` if user doesn't exist
Try twice to apply the infraction. If the user is not in the database,
try to add it, then try to apply the infraction again.

This allows any moderation function that uses `FetchedUser` as a
converter to apply the infraction even when the user is absent in the
local database.