Adding PyGame and Ren'Py to the invite white list
With the addition of the #game-development channel, I would expect we'll see an increase in people needing specific help with the various engines and libraries that Python has to offer.  As part of this, I'm adding the servers for PyGame and Ren'Py to the white list to help people get to the proper resources they might need.