Fix user command error for empty custom status
In the scenario where a user has a custom status set to only an emoji & blank text value, their activity status is set as `None`, which causes `discord.utils.escape_markdown` to raise an exception, as it's expecting something string-like.

We can guard against this by checking for a `None` value & skipping the markdown escape if necessary.