Bot: avoid DeprecationWarning for aiohttp.AsyncResolver (fix #748)
AsyncResolver has to be created inside a coroutine so it's moved
inside start(). Consequently, the APIClient session is also recreated
inside start() now.

When using clear(), the default connector is used for the session it
recreates because clear() is not a coroutine. This should only affect
requests made to the Discord API via the Client when not using it to run
a bot; starting the bot will re-create the session with the custom
connector.

* Close connector and resolver when bot closes