Error handler: simplify check failure handler & handle bot missing roles
discord.py's default error messages are quite descriptive already so
there really isn't a need to write our own. Therefore, the log calls
were removed so that the generic debug log message is used in the
on_command_error.

In addition to handling missing bot permissions, missing bot roles are
also handled. The message doesn't specify which because it doesn't
really matter to the end-user. The logs will use the default error
messages as described above, and those will contain the specific roles
or permissions that are missing.